Fix order ambiguity in weighted_order()
Also adds an option to weighted_choice() to return a 2-tuple
of (index, value) in addition to the existing option of returning
just the chosen value.

Also adds a ProbabilityUndefinedError exception to the module,
and implements it in weighted_choice() and weighted_order()

Tests have been updated to cover new behavior.